Add nested logics

+++ src/main/java/com/prisonerprice/oneAppTwoDbDemo/repository/sql/SqlUserRepository.java
@@ -3,6 +3,6 @@ package com.prisonerprice.oneAppTwoDbDemo.repository.sql;
 import com.prisonerprice.oneAppTwoDbDemo.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SqlUserRepository extends JpaRepository<User, Integer> {
+public interface SqlUserRepository extends JpaRepository<User, Long> {
      User findUserByName(String name);
 }
